docs: improve documentation, separate snapshot result list
