+++ src/lib/model/Hub.ts
@@ -19,7 +19,8 @@ import { Settings } from './Settings';
 import { Webhook, WebhooksPage } from './Webhook';
 import { WorkflowState, WorkflowStatesPage } from './WorkflowState';
 import { Extension, ExtensionsPage } from './Extension';
-import { Snapshot, SnapshotResultList } from './Snapshot';
+import { Snapshot } from './Snapshot';
+import { SnapshotResultList } from './SnapshotResultList';
 
 /**
  * Class representing the [Hub](https://amplience.com/docs/api/dynamic-content/management/#resources-hubs) resource.
