+++ src/lib/model/SnapshotCreator.ts
@@ -1,5 +1,5 @@
 /**
- * Snapshot created from
+ * Resource type a snapshot was created from.
  */
 export enum SnapshotCreator {
   ContentItem = 'content-item',
